Time: 1 ms (100.00%) | Memory: 43.9 MB (29.10%) - LeetSync

+++ 4305-check-adjacent-digit-differences/check-adjacent-digit-differences.java
@@ -4,7 +4,7 @@ class Solution {
         int n = arr.length;
 
         for(int i=0; i<n-1; i++){
-            int diff = Math.abs((int)arr[i] - (int)arr[i+1]);
+            int diff = Math.abs(arr[i] - arr[i+1]);
 
             if(diff > 2) return false;
         }
